fix(gemini): pin pricing and qualifier contracts
