+++ src/server/routes/pricing.rs
@@ -2,7 +2,7 @@
 //!
 //! This module provides HTTP endpoints for managing pricing data
 
-use crate::core::pricing_service::{LiteLLMModelInfo, PricingService};
+use crate::core::pricing_service::{LiteLLMModelInfo, PricingSnapshot};
 use crate::server::state::AppState;
 use crate::utils::error::gateway_error::Result;
 use actix_web::{HttpResponse, web};
@@ -128,8 +128,9 @@ pub async fn get_model_pricing(
 ) -> Result<HttpResponse> {
     let model_name = path.into_inner();
     let pricing_service = &data.pricing;
+    let pricing_snapshot = pricing_service.snapshot();
 
-    match effective_model_pricing_at(pricing_service, &model_name, Utc::now()) {
+    match effective_model_pricing_at(&pricing_snapshot, &model_name, Utc::now()) {
         Some(model_info) => Ok(HttpResponse::Ok().json(model_info)),
         None => Ok(HttpResponse::NotFound().json(serde_json::json!({
             "error": "Model not found",
@@ -139,12 +140,12 @@ pub async fn get_model_pricing(
 }
 
 fn effective_model_pricing_at(
-    pricing_service: &PricingService,
+    pricing_snapshot: &PricingSnapshot,
     model_name: &str,
     pricing_time: DateTime<Utc>,
 ) -> Option<LiteLLMModelInfo> {
-    let raw = pricing_service.get_model_info(model_name)?;
-    pricing_service
+    let raw = pricing_snapshot.get_model_info(model_name)?;
+    pricing_snapshot
         .get_model_info_for_provider_at(&raw.litellm_provider, model_name, pricing_time)
         .map(|(_, model_info)| model_info)
 }
@@ -276,7 +277,8 @@ mod tests {
             PricingService::with_embedded_default().expect("embedded pricing should initialize");
         let after_promotion = Utc.with_ymd_and_hms(2027, 1, 1, 0, 0, 0).unwrap();
 
-        let model_info = effective_model_pricing_at(&pricing, "gemini-3.7-flash", after_promotion)
+        let snapshot = pricing.snapshot();
+        let model_info = effective_model_pricing_at(&snapshot, "gemini-3.7-flash", after_promotion)
             .expect("Gemini 3.7 pricing should resolve");
 
         assert_eq!(model_info.input_cost_per_token, Some(1.5e-6));
@@ -285,6 +287,27 @@ mod tests {
             model_info.extra["cache_read_input_token_cost"],
             serde_json::json!(1.5e-7)
         );
+    }
+
+    #[tokio::test]
+    async fn model_pricing_resolution_stays_on_one_snapshot_generation() {
+        let pricing =
+            PricingService::with_embedded_default().expect("embedded pricing should initialize");
+        let snapshot = pricing.snapshot();
+        pricing.add_custom_model(
+            "gemini-3.7-flash".to_string(),
+            runtime_model_info("replacement_provider"),
+        );
+
+        let model_info = effective_model_pricing_at(
+            &snapshot,
+            "gemini-3.7-flash",
+            Utc.with_ymd_and_hms(2026, 8, 30, 0, 0, 0).unwrap(),
+        )
+        .expect("the pinned generation should retain Gemini pricing");
+
+        assert_eq!(model_info.litellm_provider, "vertex_ai-language-models");
+        assert_eq!(model_info.input_cost_per_token, Some(0.75e-6));
     }
 
     #[tokio::test]
